// frontend/lib/server/types.ts
export interface FilledMicroExercise {
  session_goal: string;
  quick_check_in: {
    mood_rating: number;
    primary_emotion: string;
  };
  exercise_content: {
    qna: {
      question: string;
      answer: string;
    }[];
    mcq: {
      question: string;
      options: string[];
      answers: string[];
    }[];
  };
  user_reflection: {
    mood_rating_after: number;
    reflection: string;
  };
}